+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Fixed
+- Fix typescript bindings generation for custom key and value serialization formats
+- Fix missing external messages in bindings
+
 ## [1.1.0] - 2023-04-19
 
 ## ⚡️ Breaking changes